d3d12: Fix depth buffer not being transitionned to generic read
